fix: review 도메인 구조 정리 및 후기 작성 직군 API 계약 수정
- api/types/utils 서브폴더를 review.api/hooks/lib/types.ts로 통합
- index.ts를 거치지 않던 5개 소비처의 딥 임포트를 index 경유로 수정
- Container/Stream/Content 등 범용 이름을 tag/spectrum과 통일된
  Entry/Gate/Card/Loading 역할 네이밍으로 변경
- POST /reviews의 projectPositions(배열, FE/BE/PM/PD/AI, 1~2개) 계약에
  맞게 요청 필드와 후기 작성 폼의 직군 선택 UI를 단일 선택에서
  최대 2개 다중 선택으로 수정
