+++ app/(protected)/my-soss/page.tsx
@@ -4,8 +4,7 @@ import { Suspense } from 'react';
 import { fetchMyProfile, MyProfileSectionGate, ProfileDetailView } from '@/features/profile';
 import { ProjectSectionStream } from '@/features/project/components/project-section-stream';
 import { ProjectSectionSkeleton } from '@/features/project/components/project-section.skeleton';
-import { UserReviewStream } from '@/features/review';
-import { UserReviewContainerSkeleton } from '@/features/review/components/user-review-container.skeleton';
+import { UserReviewCardEntry, UserReviewCardLoading } from '@/features/review';
 import { SpectrumCardEntry, SpectrumCardLoading } from '@/features/spectrum';
 import { TagCardEntry, TagCardLoading } from '@/features/tag';
 import { PageContainer } from '@/shared/components/page-container';
@@ -33,8 +32,8 @@ const Page = async () => {
                 <SpectrumCardEntry userLink={myProfile.userLink} showDistribution />
               </Suspense>
             </div>
-            <Suspense fallback={<UserReviewContainerSkeleton />}>
-              <UserReviewStream userId={myProfile.userId} />
+            <Suspense fallback={<UserReviewCardLoading />}>
+              <UserReviewCardEntry userId={myProfile.userId} />
             </Suspense>
           </>
         }
