+++ features/review/components/review-list/user-review-card.tsx
@@ -1,13 +1,13 @@
 'use client';
 
 import { ReviewListCard } from './review-list-card';
-import { useUserReviews } from '../../api/queries';
+import { useUserReviews } from '../../review.hooks';
 
 interface Props {
   userId: number;
 }
 
-export const UserReviewContainer = ({ userId }: Props) => {
+export const UserReviewCard = ({ userId }: Props) => {
   const { data: rawReviews } = useUserReviews(userId);
   // TODO: 백엔드 api에 projectStatus 추가 후 r.projectStatus === undefined 조건 제거
   const reviews = rawReviews.filter(
