+++ features/review/components/review-write-form.tsx
@@ -9,31 +9,20 @@ import { Button } from '@/shared/components/button';
 import { Textarea } from '@/shared/components/textarea';
 import { cn } from '@/shared/lib/cn';
 
-import type { UserPosition } from '../types/review';
+import type { ReviewPosition } from '../review.types';
 
 import { ReviewSpectrumRow, spectrumStepToValue } from './review-spectrum-row';
 import { ReviewSubmitDialog } from './review-submit-dialog';
-import { useCreateReview } from '../api/mutations';
-import { useReviewFormData } from '../api/queries';
+import { useCreateReview, useReviewFormData } from '../review.hooks';
+import { REVIEW_POSITIONS, REVIEW_POSITION_LABELS } from '../review.lib';
 
 const FEEDBACK_MIN_LENGTH = 10;
 const FEEDBACK_MAX_LENGTH = 1000;
 const MAX_TAGS = 3;
+const MAX_POSITIONS = 2;
 const DEFAULT_SPECTRUM_STEP = 2;
 
-const USER_POSITION_LABELS: Record<UserPosition, string> = {
-  FRONTEND: '프론트엔드',
-  BACKEND: '백엔드',
-  PM: 'PM',
-  PD: 'PD',
-  AI: 'AI',
-  QA: 'QA',
-  ETC: '기타',
-};
-
-const USER_POSITIONS = Object.keys(USER_POSITION_LABELS) as UserPosition[];
-
-export const ReviewWriteContent = () => {
+export const ReviewWriteForm = () => {
   const router = useRouter();
   const searchParams = useSearchParams();
   const projectId = useMemo(() => {
@@ -64,8 +53,7 @@ export const ReviewWriteContent = () => {
   const [selectedTagIds, setSelectedTagIds] = useState<Set<number>>(() => new Set());
   const [spectrumSteps, setSpectrumSteps] = useState<Record<number, number>>({});
   const [feedback, setFeedback] = useState('');
-  const [projectPosition, setProjectPosition] = useState<UserPosition | null>(null);
-  const [projectDetailPosition, setProjectDetailPosition] = useState('');
+  const [selectedPositions, setSelectedPositions] = useState<Set<ReviewPosition>>(() => new Set());
   const [submitDialogOpen, setSubmitDialogOpen] = useState(false);
   const [submitError, setSubmitError] = useState<string | null>(null);
 
@@ -84,23 +72,39 @@ export const ReviewWriteContent = () => {
     });
   }, []);
 
+  const togglePosition = useCallback((position: ReviewPosition) => {
+    setSelectedPositions((prev) => {
+      const next = new Set(prev);
+      if (next.has(position)) {
+        next.delete(position);
+        return next;
+      }
+      if (next.size >= MAX_POSITIONS) {
+        return prev;
+      }
+      next.add(position);
+      return next;
+    });
+  }, []);
+
   const onSpectrumChange = useCallback((spectrumId: number, step: number) => {
     setSpectrumSteps((prev) => ({ ...prev, [spectrumId]: step }));
   }, []);
 
   const feedbackTrimmed = feedback.trim();
   const feedbackOk = feedbackTrimmed.length >= FEEDBACK_MIN_LENGTH;
-  const positionOk = projectPosition != null && (projectPosition !== 'ETC' || projectDetailPosition.trim().length > 0);
+  const positionOk = selectedPositions.size > 0 && selectedPositions.size <= MAX_POSITIONS;
   const tagsOk = selectedTagIds.size > 0 && selectedTagIds.size <= MAX_TAGS;
   const spectrumsOk = !!formData?.spectrums?.length;
   const canSubmit = feedbackOk && positionOk && tagsOk && spectrumsOk && !isSubmitting;
 
   const handleSubmitFromDialog = useCallback(async () => {
-    if (!formData || !canSubmit || projectId == null || revieweeId == null || projectPosition == null) {
+    if (!formData || !canSubmit || projectId == null || revieweeId == null) {
       return;
     }
     setSubmitError(null);
     const tagIds = [...selectedTagIds];
+    const projectPositions = [...selectedPositions];
     const spectrums = hasSubmittedAnyReview
       ? []
       : formData.spectrums.map((s) => ({
@@ -112,8 +116,7 @@ export const ReviewWriteContent = () => {
         projectId,
         revieweeId,
         feedback: feedbackTrimmed,
-        projectPosition,
-        projectDetailPosition: projectPosition === 'ETC' ? projectDetailPosition.trim() : undefined,
+        projectPositions,
         tagIds,
         spectrums,
       });
@@ -127,12 +130,11 @@ export const ReviewWriteContent = () => {
     canSubmit,
     hasSubmittedAnyReview,
     selectedTagIds,
+    selectedPositions,
     spectrumSteps,
     projectId,
     revieweeId,
     feedbackTrimmed,
-    projectPosition,
-    projectDetailPosition,
     submitReview,
     router,
   ]);
@@ -246,41 +248,36 @@ export const ReviewWriteContent = () => {
 
           <section className="flex flex-col gap-4" aria-labelledby="review-position-heading">
             <h2 id="review-position-heading" className="text-heading-sm text-text-basic leading-normal font-bold">
-              프로젝트에서의 직군
+              프로젝트에서의 직군(최대 {MAX_POSITIONS}개)
             </h2>
             <p className="text-body-sm text-text-subtle -mt-2">(필수) 이 프로젝트에서 맡은 역할을 선택해주세요.</p>
             <div className="flex max-w-[781px] flex-wrap gap-x-2 gap-y-2">
-              {USER_POSITIONS.map((pos) => (
-                <button
-                  key={pos}
-                  type="button"
-                  aria-pressed={projectPosition === pos}
-                  className={cn(
-                    'text-body-sm inline-flex h-[33px] shrink-0 items-center justify-center rounded-full border px-3 font-normal transition-colors outline-none focus-visible:ring-2 focus-visible:ring-(--color-border-primary) focus-visible:ring-offset-1',
-                    projectPosition === pos
-                      ? 'border-border-gray-light bg-action-secondary-pressed text-text-basic'
-                      : 'border-border-gray-light bg-action-gray-light text-text-basic hover:border-action-secondary-hover hover:bg-action-secondary-hover',
-                  )}
-                  onClick={() => {
-                    setProjectPosition(pos);
-                    if (pos !== 'ETC') {
-                      setProjectDetailPosition('');
-                    }
-                  }}
-                >
-                  {USER_POSITION_LABELS[pos]}
-                </button>
-              ))}
+              {REVIEW_POSITIONS.map((pos) => {
+                const selected = selectedPositions.has(pos);
+                return (
+                  <button
+                    key={pos}
+                    type="button"
+                    aria-pressed={selected}
+                    disabled={!selected && selectedPositions.size >= MAX_POSITIONS}
+                    className={cn(
+                      'text-body-sm inline-flex h-[33px] shrink-0 items-center justify-center rounded-full border px-3 font-normal transition-colors outline-none focus-visible:ring-2 focus-visible:ring-(--color-border-primary) focus-visible:ring-offset-1',
+                      selected
+                        ? 'border-border-gray-light bg-action-secondary-pressed text-text-basic'
+                        : 'border-border-gray-light bg-action-gray-light text-text-basic hover:border-action-secondary-hover hover:bg-action-secondary-hover',
+                      !selected &&
+                        selectedPositions.size >= MAX_POSITIONS &&
+                        'hover:border-border-gray-light hover:bg-action-gray-light cursor-not-allowed opacity-30',
+                    )}
+                    onClick={() => {
+                      togglePosition(pos);
+                    }}
+                  >
+                    {REVIEW_POSITION_LABELS[pos]}
+                  </button>
+                );
+              })}
             </div>
-            {projectPosition === 'ETC' && (
-              <input
-                type="text"
-                placeholder="직군을 직접 입력해주세요"
-                value={projectDetailPosition}
-                onChange={(e) => setProjectDetailPosition(e.target.value)}
-                className="border-border-gray-light text-body-sm text-text-basic placeholder:text-text-disabled mt-1 max-w-[400px] rounded-md border px-3 py-2 outline-none focus:ring-2 focus:ring-(--color-border-primary)"
-              />
-            )}
           </section>
         </div>
