+++ features/review/review.api.ts
@@ -1,18 +1,44 @@
 import { apiRequest } from '@/shared/lib/api';
 import type { ApiRequestOptions } from '@/shared/lib/api';
 
-import type { CreateReviewRequest, Review, ReviewFormData, ReviewValidation } from '../types/review';
+import type { ReviewApiRaw, ReviewFormDataApiResponse, UserReviewsApiResponse } from './review.lib';
+import type { CreateReviewRequest, Review, ReviewFormData, ReviewPosition, ReviewValidation } from './review.types';
 
-import { mapReviewFormDataFromApi, type ReviewFormDataApiResponse } from './map-form-data';
-import { mapUserReviewsFromApi, type ReviewApiRaw, type UserReviewsApiResponse } from './map-user-reviews';
+import { mapReviewFormDataFromApi, mapUserReviewsFromApi } from './review.lib';
 
-const toCreateReviewBody = (data: CreateReviewRequest) => ({
+export const reviewKeys = {
+  all: ['review'] as const,
+  reviews: (userId: number) => [...reviewKeys.all, 'reviews', userId] as const,
+  formData: () => [...reviewKeys.all, 'formData'] as const,
+  projectReviews: (userId: number, projectId: number) => [...reviewKeys.all, 'reviews', userId, projectId] as const,
+  validate: (projectId: number, revieweeId: number) => [...reviewKeys.all, 'validate', projectId, revieweeId] as const,
+};
+
+interface CreateReviewReviewReqDto {
+  projectId: number;
+  revieweeId: number;
+  feedback: string;
+  projectPositions: ReviewPosition[];
+  tagIds: number[];
+}
+
+interface CreateReviewSpectrumReqDto {
+  spectrumAxisId: number;
+  spectrumStrength: number;
+}
+
+/** POST /api/v1/reviews 요청 본문 (백엔드 ReviewCreateReqDto) */
+interface CreateReviewApiBody {
+  reviewReqDto: CreateReviewReviewReqDto;
+  spectrumReqDtos: CreateReviewSpectrumReqDto[];
+}
+
+const toCreateReviewBody = (data: CreateReviewRequest): CreateReviewApiBody => ({
   reviewReqDto: {
     projectId: data.projectId,
     revieweeId: data.revieweeId,
     feedback: data.feedback.trim(),
-    projectPosition: data.projectPosition,
-    ...(data.projectDetailPosition ? { projectDetailPosition: data.projectDetailPosition } : {}),
+    projectPositions: data.projectPositions,
     tagIds: data.tagIds,
   },
   spectrumReqDtos: data.spectrums.map((spectrum) => ({
