--- features/review/types/spectrum.ts
@@ -1,10 +1,0 @@
-export interface Spectrum {
-  spectrumId: number;
-  leftLabel: string;
-  rightLabel: string;
-}
-
-export interface SpectrumWithValue {
-  spectrumId: number;
-  value: number;
-}
